ch: rename RefreshButton to RefreshMessage
